refactor(markdown): share iframe sanitizer
