+++ packages/plugins/@nocobase/plugin-field-markdown-vditor/src/client-v2/components/Display.tsx
@@ -7,11 +7,11 @@
  * For more information, please refer to: https://www.nocobase.com/agreement.
  */
 
+import { removeMarkdownIframes, stripMarkdownIframes } from '@nocobase/client-v2';
 import { Popover } from 'antd';
 import React, { CSSProperties, useCallback, useEffect, useRef, useState } from 'react';
 import Vditor from 'vditor';
 import { useCDN } from './const';
-import { removeMarkdownIframes, stripMarkdownIframes } from './sanitize';
 import useStyle from './style';
 
 function convertToText(markdownText: string) {
